requirements-itest.txt
argparse
googleapis-common-protos
grpcio
grpcio-tools
importlib_resources==1.4.0
pytest
squeaklib==0.5.1
